+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+...
+
+## [0.3.0] — 2019-02-09
 - **Breaking change.** Rename `DjangoModelAutoFactory.Meta.fields` into
 `DjangoModelAutoFactory.Meta.autofields`.
 - Add `DjangoModelAutoFactory.Meta.autoexclude`.
@@ -20,5 +23,6 @@ value.
 ## 0.1.0 — 2019-02-03
 - Add Django ORM support.
 
-[Unreleased]: https://github.com/nickgashkov/autofactoryboy/compare/v0.2.0...HEAD
+[Unreleased]: https://github.com/nickgashkov/autofactoryboy/compare/v0.3.0...HEAD
+[0.2.0]: https://github.com/nickgashkov/autofactoryboy/compare/v0.3.0...v0.2.0
 [0.2.0]: https://github.com/nickgashkov/autofactoryboy/compare/v0.2.0...v0.1.0